+++ src/api/auth.rs
@@ -23,17 +23,6 @@ where
     U: Display,
     P: Display,
 {
-    /// Returns the current OAuth 2.0 configuration for authentication.
-    /// See [OAuth 2 Guide](https://www.rabbitmq.com/docs/oauth2) to learn more.
-    ///
-    /// Requires the `management` user tag. Does not modify state.
-    pub async fn oauth_configuration(&self) -> Result<responses::OAuthConfiguration> {
-        let response = self.http_get("auth", None, None).await?;
-        let response = response.json().await?;
-
-        Ok(response)
-    }
-
     /// Returns authentication attempt statistics for a given node.
     ///
     /// Requires the `monitoring` user tag. Does not modify state.